Updated openstack/openstack Project: openstack-dev/devstack  8cdeb866481d91316c9fc0765b15da894a5a278a
lib/neutron-plugins: Pass conf file parameter

The 'neutron_plugin_configure_l3_agent' function expects a path to a
configuration file as a parameter. This was not done for one call,
resulting in the generation of a 'DEFAULT' file in the DevStack
directory along with an invalid L3 configuration file. Resolve this.

Change-Id: I5781cb1ec4cfc1699e61dbc324d0bdb824b56be1