Updated openstack/openstack Project: openstack/nova  9a26e9236dd9aeedd884428e62fb986716ec17b4
XenAPI: Integers returned from XAPI are actually strings

In particular, we were expecting the domid to be an integer
when comparing against -1 in vmops.py to decide if the domain
was shut down.  This test always failed, so domid of -1 was
assumed to be a live domain.

Fix all places where domid is an integer to ensure only string
comparisons are done.
If the domid is '-1' then we raise an exception based on the
instance.  An instance is always passed, so make it clear that
this is non-optional

Change-Id: I8c0d3274fbf68351a1b077e791e50e449ea0af99
Closes-Bug: 1592849