Updated openstack/openstack Project: openstack/neutron  dab8083542fc6fec2eb6183b8f7c4438ed85a8ef
ovs: set device MTU after it's moved into a namespace

Open vSwitch forbids to raise MTU for devices that are in the same
namespace with the bridge, but allows that same operation once the
device in question is moved into another namespace. This is a side
effect of missing validation in vswitchd implementation for plugged
interfaces located in other namespaces.

This behaviour makes the order in which we move device into a namespace
and set MTU significant. The previous order left MTU lowered for router
and dhcp ports in case there was a port served by that same bridge with
lower MTU.

This effectively lowers MTU for router interfaces from what is supported
by underlying physical interfaces. With that, frames from an external
network of infrastructure MTU size will not be served by affected
routers.

This Neutron issue affects all systems where router or dhcp ports are
served on the same node as compute ports. This includes any AIO setup,
as well as L3/DHCP agents colocated with nova-compute. While for
traditional (legacy) L3 agents, such setup is not that widely used, the
bug probably affects real world DVR setups that maintain router
namespaces on compute nodes.

The fix reverses the order of interface plugging procedure steps, first
moving ports into namespaces, then setting their MTUs. This seems to
solve the issue.

The fix is short term, because we effectively utilize a bug in Open
vSwitch. A better long term fix could be rearranging bridge setup so
that there is no need to plug devices that belong to different networks
into the same bridges. That would require an extensive thought and would
open questions on how to migrate existing workloads using the current
bridge setup to the new world without significant data path disruption.

Added functional test that triggered the issue before the fix.

Change-Id: Ibd9866d2d8fe12bb253c6fb9978f583293b32734
Closes-Bug: #1590397