Updated openstack/openstack Project: openstack/ironic  4dd33f8bd48df1df58eaac03872b133cc914bd8e
Replace dict.get(key) in drivers/modules/*/ tests

This replaces dict.get(key) with dict[key] in
driver/modules/<vendor> unit tests, since we should
be explicit about whether we expect the key to exist or not.

In particular, we shouldn't be doing
    self.assertIsNone(dict.get(key))
since it will pass if 1. the value is None or 2. if the key isn't
in the dict.

Change-Id: I1d0bc7d5b5561eab4317e47f2d7ef006f9e1cb1f
Partial-Bug: #1590808