Updated openstack/openstack Project: openstack/cinder  1af0da7876703811a32f8f9e99040236ea1c4561
use utf8 as default charset for all tables in mysql db

When we use "cinder-manager db sync" to create tables in mysql db,
some tables' charset is set to utf8 but some not. This will cause
db error when a foreign key is referenced to column in different
charset. This change set all tables in cinder use utf8 as default
charset.

Closes-bug: #1594195

Change-Id: I15a864dd1ad18aa1b3e3ff7fe511405d62a119cc