Updated openstack/openstack Project: openstack-infra/project-config  e29eaacb6767ca39d3339e81950bf3f39b52173e
Promote existing OSA non-voting Xenial tests to voting

After a few weeks of testing as non-voting jobs, OpenStack-Ansible
roles which have had non-voting jobs are being promoted to voting.

The pip_lock_down role is deliberately left out as it is scheduled
for retirement within this cycle and will not require Xenial
support.

Change-Id: Ia884f90d0d48578bbce81d2b48db47fee3b9098d