Updated openstack/openstack Project: openstack-infra/storyboard-webclient  f10e59129118b621baf9f4cd31107952b08fb682
Add the ability to permalink to a comment

This commit adds a button to comments which directs you to a link that
will automatically scroll the page to the comment.

The link can be manually constructed by appending `#comment-$ID` to the
story's URL. For example, to link to a comment with ID 45, one would
append `#comment-45`.

Change-Id: Iacf43c81457b018a30a2f1a410cfa50bea972fef