Updated openstack/openstack Project: openstack/cinder  5447efe7b5b679c60c7e3e839201bc50249ec240
Minimize cost of DB exists decorator checks

We have 2 DB decorators, require_volume_exists and
require_snapshot_exists, that rely on retrieving the resource from the
DB to make sure that it really exists, but we don't really need to
retrieve the data from the DB and build the ORM model from the data in
order to decide if the resource exists.

This patch modifies both decorators to use a more efficient method using
EXISTS(SELECT 1 FROM ... WHERE ...) which not only reduces the amount of
data retrieved from the DB but also removes the parsing of the data into
the resource's ORM model.

Change-Id: I90dec6dda5b2c91b78c1257afdb1b749f4bdc030