Updated openstack/openstack Project: openstack/ironic  b83040bd3045448c6b7cd9d324a5728b2df17fdd
Replace dict.get(key) in drivers unit tests

This replaces dict.get(key) with dict[key] in
tests/unit/drivers & tests/unit/drivers/modules, since we should
be explicit about whether we expect the key to exist or not.

In particular, we shouldn't be doing
    self.assertIsNone(dict.get(key))
since it will pass if 1. the value is None or 2. the key isn't
in the dict.

Change-Id: I47388b2f8650f2110755129da63dc7f482a71bbf
Partial-Bug: #1590808