Updated openstack/openstack Project: openstack/horizon  8120b7d1e097ef77cde4a5688a8b510af80386ed
Relocated error message to associated field and corrected grammar

Validation error was displaying at the top of the form making it
seem as though it was a redundant error. Added validation so that
error appears below the affected field. Fixed the grammar of the
original error message (“a image” -> “an image”).

Change-Id: I7250a44138a3559ddf39ff9d7bde1478ecb3a143
Closes-Bug: 1570510