Updated openstack/openstack Project: openstack/neutron  b59bb0fcfa41963c0e2f7bcbf34b7e4f4ff5cd08
Obsolete mac_generation_retries and deprecate the option

As discussed in this patch [1] this option didn't come from any actual
user-encountered problem and is now effectively obsolete for its
original purpose anyway.  In the interest of cleaning up options that
aren't of any real use, this patch obsoletes this one immediately and
marks it deprecated for removal in Ocata.

TrivalFix

[1] https://review.openstack.org/#/c/327413

Change-Id: I58003d5b07e28c7254ff836dec937f699bc83112