Updated openstack/openstack Project: openstack/barbican  c6342039cc370ab42f2e0254d307c8ee950cfd9a
Setup memory DB in test_cmd

Currently, test_cmd fails or not depending if
database_utils.setup_in_memory_db() was already called. It looks like
the test doesn't fail on Python 2, but it fails *sometimes* on
Python 3.

Maybe the tests are run in a diffrent order on Python 3 because of
the randomized hash function (it can be tried on Python 2 using -R
Python command line option).

The bug can been reproduced easily be running only test_cmd
on Python 2.7:

    $ . .tox/py27/bin/activate
    $ testr run barbican.tests.cmd.test_cmd
    ...
    FAIL: barbican.tests.cmd.test_cmd.WhenInvokingWorkerCommand.test_should_launch_service
    ...
    BarbicanException: No SQL connection configured

Change-Id: Ic68582193fa4a5002cab8333797eb3fe2881f051