Updated openstack/openstack Project: openstack/neutron  7245b172b0df4cc0a6b062f9bebcc062aa950250
How to support trunk ports with Open vSwitch Agent

This patch documents the various strategies that can be
considered to support the vlan-aware-vms API with the
OVS mech driver and respective agent.

A proposal is chosen based on the tradeoffs amongst design,
implementation, upgrade, and performance complexities, and
it has the aim of fostering brainstorming and discussion.

More patches will follow up on design internals for the
adopted implementation strategy.

Partially-implements: blueprint vlan-aware-vms

Co-Authored-By: Armando Migliaccio <armamig@gmail.com>
Co-Authored-By: Rossella Sblendido <rsblendido@suse.com>

Change-Id: Ib4a176da2863b796283d6c1fff05f08edf40efb4