Updated openstack/openstack Project: openstack/python-openstackclient  b86915aa7e2152b746d2c71c81b4c67a0ba9490e
Refactor check_valid_auth_options function

The functions check_valid_auth_options() function was relying on the name
for checking the set of required options, but this could cause errors
with external auth plugins. If somebody defines an auth plugin plugin
named "footoken" the check function would check for a "token" option,
even if the plugin has not defined that option. This change tries to
improve this situation, cheking for some options only if they have been
defined in the plugin.

Change-Id: I4255f2e7d4d23449c95be957ea7b6b60983f2608