Updated openstack/openstack Project: openstack/keystone  8f8f85fbfc8485791fe7ea537efb98453eb9b499
Correct domain_id and name constraint dropping

The 'domain_id' and 'name' unique constraint was not properly dropped
in some cases because the unique constraint was not consistently
named. In all cases we must search for the constraint expected,
not assume the name of the constraint will be consistent
(especially from older installs that have been moved forward in
releases). This fix is modeled on the fix for a similair issue
authored by Morgan Fainberg & Matthew Thode for Bug #1562934

Migration 091:
    Fix to broken migration to prevent failed migrations when database is
    upgraded from Kilo (or below) to Mitaka
Migration 097:
    Ensure that when Mitaka point release is applied the constraint and tables
    have been dropped if migration 91 was previously worked around.
Migration 104:
    Ensure that when upgrading to Newton the constraint and tables
    have been dropped if migration 91 was previously worked around.

Migration 91 drops 3 columns from the user table after the code to disable
the constraint. I have included code in migrations 97 and 104 to also drop
those columns if they are still present in case they were missed when working
around Bug #1572341. This may be over kill.

Change-Id: I076d7139b388e30be8826d0a4550256b5617d992
Closes-bug: #1572341