Updated openstack/openstack Project: openstack/keystone  2088906a04d0f337a721503d6aa69fe2d3c6f1a9
Make sure to use InnoDB as the DB engine

For certain Fedora based systems the MySQL default DB engine is MyISAM.
Therefore, the MySQLOpportunisticUpgradeTestCase functional test suite
will fail because of MyISAM's known limitations. We've already established
a pattern to mitigate this problem, which is to simply force the upgrade
scripts to use InnoDB engine explicitly.

Change-Id: Idd20ea7f0f0a41c01d62b8730c7bd4ed6ecd6486
Closes-Bug: 1594592