Updated openstack/openstack Project: openstack-infra/zuul  0bcf236cf1eeac76174e10b060caf141d5f125cf
Ansible launcher: copy lib/plugins at startup

Because we run Zuul continuously deployed from source and installed
via pip, whenever a Zuul commit lands, pip re-installs Zuul.  When
that happens, there may be several minutes where part of the Zuul
source code is no longer on disk.  Because the ansible launcher
points directly into the installed Zuul source for the zuul library
and plugins, jobs launched during that window will fail.  To correct
this, make a copy of the necessary ansible files on startup and point
the jobs to that location.

Change-Id: Ib0b6f07967ba314d34fad4835c4e2e84584e446c