Updated openstack/openstack Project: openstack/tripleo-heat-templates  a2614491be1cae82e4b04337184b75644633ce3f
Composable roles within services - NTP

Add NTP as a composable service.

Also, the scope of this patch is to serve as
basis for the blueprint related to the
composable services walkthrough.

This will be the use case documented.

The puppet manifests updated to remove the current
configuration for NTP is also added.

Change-Id: I1d3d6229d53f48c380e6a775a7c50de8715fe949