Updated openstack/openstack Project: openstack/neutron  e7d12dadbc8a6623e31f20e94a1e7ca5bf04348b
Disable warnerrors=True in setup.cfg

This feature is not working for a while, which potentially allowed
errors to sneak into the tree. The next pbr release will fix the
feature, effectively re-enabling it for us. It may break our gate if at
that point we will have warnings in the documentation.

So just disable the flag for now, wait for the next pbr release, and
revert the patch back, fixing any warnings that may be present in the
tree at that time.

Change-Id: I117a458d2b4b043fa0238eebc061586761691841