Updated openstack/openstack Project: openstack/horizon  2620791889f2d9604d6d50de88f20bc522ae3dc3
Add check whether nova is enable or not in aggregate panel

At the moment, aggregate panel has a check whether nova extension
is enable or not. However it doesn't check nova is enable or not.
So, when operator use non nova, error always occur by this check
method. Therefore, this patch will add a check whether nova is
enable or not.

Closes-Bug: #1591920

Change-Id: I5ad37f16f70f1ac9b95c49b1363d1b23abbb1b94