Updated openstack/openstack Project: openstack/glance_store  6f4307cccff65fa8c50ee6040f5a3d6d02dd977a
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed). If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit
Tests fail. The following patch fixes this issue.

TrivialFix

Change-Id: I07b78383ef38731140143c91ae3a902bea55eebb
Closes-Bug: #1259292