Updated openstack/openstack Project: openstack/heat  59b9c3adaf7953ce0b45aab1e8c564b8da513e8d
Improve unit tests for stack-update-cancel

There's little point in having a unit test that mirrors the logic of the
function it is testing, so refactor to assert the behaviour we want
explicitly. Also, add a test for a missing case.

In addition, the test_update_rollback_on_cancel_event test should have
changed the message from 'cancel' to 'cancel_with_rollback' in the patch
b76f6ec0e97a9c3aaf5b5140e7a449c9636b8f95, but this was inadvertently
omitted.

Change-Id: I70e51255ffd1e7f92d060c8da075f4ce4c506079