Updated openstack/openstack Project: openstack/ceilometer  c51c9f1564bb2b74474eca5578ce2dd23136b531
Merge tag '6.1.0'

This is a null-merge of the 6.1.0 release tag back into the master
branch so that the 6.1.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I94019769a3ebcaba66e8bc1544a3563f325c1735