Updated openstack/openstack Project: openstack/nova  9e0d98eb9805ca79fb4b692e643f3a4c9f382fcc
Skip network allocation if 'none' is requested

If we're explicitly told to not allocate networks, then
don't, skip network allocation.

This also skips network deallocation if we can tell that
we wouldn't have allocated networks because 'none' was
requested for networking. Note, however, that this is just
an optimization when requested_networks is available - which
is not the case with a local delete from the compute API.
Even if we didn't check in the deallocate case it would be fine
since deallocate_for_instance in nova-network and neutron
network APIs won't find any resources to delete since none
were allocated to begin with. But at least we can try to save
ourselves some RPC/DB (nova-network) and REST API (neutron)
traffic.

Part of blueprint get-me-a-network

Change-Id: Ia0d5189453df825805dab028cc81de7dfa0d6c84