Updated openstack/openstack Project: openstack/tempest  5903016a92a46ae804bd78dae1acf015da9f1561
Make `tempest init` working dir an abspath

The working directory can be either a relative path or absolute path
depending on how `tempest init` is called. If `tempest init` is called,
the working directory is absolute, but if `tempest init [dir]` is
called, the working directory is relative. The relative case leads to a
bug in setting up the tempest config file later.

We normalized and absolutized local_dir early on, and make everything
work with abspath. This minimize handling of the different paths and
reduces the chances of bugs from handling in later code.

Change-Id: I6dc314017eb264302ba68bbf25060e833d8e317a
Closes-Bug: #1593747