Updated openstack/openstack Project: openstack/tripleo-heat-templates  203e9256f4a26696589a78f449765550960987fa
Heat upstream default policy.json works in devstack

It does not however work in TripleO, and the ping test is failing
with master Heat because of this. I tried submitting a revert[1] in
Heat, but they have functional tests, so it is not broken...

The only option left is using puppet-heat to override the broken
default.

[1] https://review.openstack.org/#/c/332206/

Change-Id: I2c674a2e01eeacd17d3158f378be2d6ceeedc309