Updated openstack/openstack Project: openstack/trove  3d1382021beee900f0ebe7025980e337c960b394
Delete with fail nova instance tries to stop DB

When deleting an instance in ERROR state, there is an attempt to stop the
database which can time out and take 5 minutes. To prevent the timeout add
check to see if instance is in ACTIVE/SHUTDOWN state before attempting to stop
the database.

Change-Id: Ie83a069a57bf13494a457c41b0af50ba235ff736
Closes-Bug: #1589676