Updated openstack/openstack Project: openstack/neutron  61a9b362130812687725808038bf0dfb1758e2d0
Remove IP availability range recalculation logic

This patch removes unreachable code that rebuilds IP address
availability data and persists in the database. This is
ultimately derived data that is now computed in memory and never
persisted. The code being removed is dead code and does not
include the contract migration for removal of the
IPAvailabilityRange models.

Change-Id: I96cb67396b8e0ebbe7f98353fad1607405944e44
Partial-Bug: #1543094