Updated openstack/openstack Project: openstack/glance  f0e5280f26f59f32252d752d529e555692520ca1
Do not set header if checksum doesn't exist

When image caching is disabled, we don't set the 'Content-MD5' header
if there is no checksum. We need to do the same when cache middleware
is enabled. This will make the response consistent. Second, it also
fixes an issue where we try to encode the header without checking
if it's defined.

Closes-Bug: #1591004

Change-Id: I3b3c58db3a839728a0d019594b2cecce4f0f8a1e