Updated openstack/openstack Project: openstack/tripleo-heat-templates  1a481a89a359e90baa0accfc6b99c8c34bfd2e1f
Drop extraconfig for neutron-opencontrail.yaml

This patch drops the extraconfig interface in favor
of using the composable services nested stack instead.

The benefit is that it is easier to enable multiple services
(like network and storage backends at the same time) and all
of the opencontrail settings get to live in the same file.

Partially-implements: blueprint composable-services-within-roles

Change-Id: I0edbd86a8c981bd6e8a547cd2a6ebed18ecdbb31