Updated openstack/openstack Project: openstack/python-openstackclient  51fcd7c30c466b1241b6f148390c83caca578c49
Refactor unit test of "compute service list" command

The unit test of "compute service list" only checked the
"Disabled Reason" columns and its data. It is not enough.
This patch change the test to check all datas in the list.

Also, this patch modify the "Id" to "ID" in this command.

Change-Id: I988fd6365f2652185dd96d9417f294eba9c31cd9