Updated openstack/openstack Project: openstack-infra/storyboard-webclient  d03592b2dd5224cbb1ba1798bdea3911ff2c218c
Fix alignment of long-running simple events

In the timeline, when simple events text runs on multiple lines,
the box moves to the left, which makes it look ugly.

See http://imgur.com/a/jElJi for an example.

Borrow CSS from the comment event to properly display the event
in such cases.

Change-Id: Iba60d19ff78c45120cceb4363260097ed3c4c4f5