Updated openstack/openstack Project: openstack/tempest  e0ef05ce0dfcfc2e2c7c4b9524af7a67bd52a1c3
Move oslo config generator config inside package

This commit moves the oslo config generator config file from living as
a data file, which makes it very difficult to rely on, to just living
in the package. By including it in the package we have a guaranteed
relative path we can rely on for building utilities on it, like the
tempest init command.

Change-Id: I5b9deae733f42fe3ed39418103304f71b43dccb3
Partially-implements: bp tempest-run-cmd