Updated openstack/openstack Project: openstack/nova  6a119fc72096fa65678112e151e4a7a00091ce26
XenAPI: OVS agent updates the wrong port with Neutron

With xen hypervisor, it will create two sets of port/interface
for each VM vif: One is named as tapx.y which is qemu-emulated
NIC and is used when no PV drivers installed with the VM; The
other one is named as vifx.0 which is the xen network frontend
NIC and is used when VM has PV drivers installed. But neutron
only handle one port. If the finally used port is not the one
handled by neutron, the guest VM will have no network access.
The solution is to change nova xenapi to add a bridge, which
we call the 'interim bridge', between the VM and the OVS
integration bridge. In this way it will only expose to neutron
with the integration bridge's port which connected to the
interim bridge. So it will fix the issue mentioned above.

Closes-Bug: #1268955

Change-Id: I0cfc0284e1fcd1a6169d31a7ad410716037e5cc2