Updated openstack/openstack Project: openstack/governance  c48d9a1528b4fa119190fa3577701daeebfbf362
Preserve superseded resolutions

Rather than directly amending a previously-passed resolution, the TC
has decided the change should be a new resolution.

For the stackforge resolution, this commit adds the text of the
updated resolution as a new resolution with some introductory
explanatory text noting what the update is and the reason for it.
It also reverts the text of the original resolution to that which
was originally passed and adds a new section at the bottom of the
original resolution which indicates it is amended by a subsequent
resolution and links to it.

The superseded resolutions are placed in a specific folder for
clarity. The same process was applied to the Mission statement
resolution.

Change-Id: I446d0797554311b199fce999a7469dbf4e3a2d30