Updated openstack/openstack Project: openstack/nova  c8ffe59b126cfa9cd49be8e0f43964eb7cefc994
Add ironic mac address when updating and creating

We should ensure we add a mac_address from the pool of available
addresses (for ironic) both when the user has already created a port and
when Nova is creating a port for the user.

This patch includes catching the MacAddressInUseClient exception when
updating a port that has possibly updated the mac. This is to match the
behavior when we call create_port.

The data for  mac addresses comes from the virt driver, so eventually
this will only happen during port update. As a first step we move it to
be called during both create and update paths.

blueprint prep-for-network-aware-scheduling

Change-Id: I536f65186aa21255dee9b1b784f1e7b1324c5cbe