Updated openstack/openstack Project: openstack/glance  df5a4552f417eac5a21c3c9fdcda521b1e374faf
Fix bug Swift ACL which disappears on Glance v1 images

Currently, when we make an update on a public image, Swift ACL disappears.
This patch allows to avoid the ACL update if there is no change to the
is_public_attribute.
If an update is made on a Glance image without changing the public attribute
or location, then the function which modifies ACLS is not called, in order
to keep ACLS.

Closes-Bug: 1584076

Change-Id: I21e9deb9105faee87500998eff47c6faa1d02914