Updated openstack/openstack Project: openstack/nova  e5eb40732d443e870db905e1091133b51a340761
Handle UnableToAutoAllocateNetwork in _build_and_run_instance

The network API will raise UnableToAutoAllocateNetwork if,
well, it's unable to automatically allocate a network for the
project (if one is not already available or requested).

For Neutron this most likely means something has gone wrong
in the deployment (since we checked a dry-run that we could do
this in the API) and rescheduling isn't going to fix it, so we
want to abort the server build and put it into ERROR state.

Part of blueprint get-me-a-network

Change-Id: I2b04ef162192c0448e881c505b15d28ba585be19