Updated openstack/openstack Project: openstack/horizon  fbb6f6411410a53bef5c49273fdf77f8c98f7eb7
Add Angular Project Volumes Panel

This patch creates a blank Volumes panel in the project dashboard.

It follows the pattern established by the following merged commits:
https://review.openstack.org/#/c/200324/

To be added in subsequent patches:
 - Actions
 - Tables
 - Filters via Magic Search
 - Integration to Searchlight

To test:
It doesn't work with only this one.
You can test volumes panel and table with this patch.
https://review.openstack.org/#/c/319051/

Partially-Implements: blueprint angularize-volumes-table
Change-Id: I17fdf46162187cd127d645370673fe3a3fe39db4