Updated openstack/openstack Project: openstack-infra/storyboard  72d4a8617925655038c4004af2f26c5e3689de8e
Allow users to be filtered by openid and email

This will allow matching of Gerrit users to StoryBoard users
to be done more easily and reliably.

These are filtered only by exact matches to reduce the ease of
brute-forcing the email addresses we keep private, and because
there is no need to do substring matching on openids.

Change-Id: Ibf824868cdbd2e654e1b108bb933dceddc53dcbf