Updated openstack/openstack Project: openstack/tempest  d23af14447c5b0ceecfd4c560a5a7050f4ace4c1
Remove testscenarios usage from test_server_basic_ops

This commit removes the scenario test wrapper around
test_server_basic_ops. This has always caused problems in tempest as
it actually tries to do real operations during discovery which is a
terrible idea. This weighed against a very limited benefit and limited
use by most people it's time to just call the experiment and remove
it. This is not saying there is no usefulness in testing a full matrix
of images and flavors on a real deployment, but instead that the place
for this kind of testing is not in tempest discovery. If people who
found this valuable in the past want to retain the functionality should
help work on an alternative method of doing it that isn't so invasive.

Change-Id: Ie7814bd50f1c57c5f8da8355f665f09c9ac8e2d9